fix: moved mock to models folder
